+++ src/autoComplete/AutoComplete.tsx
@@ -9,6 +9,8 @@ import { ItemStatus } from '_utils/item'
 import ItemInfo from 'itemInfo'
 import TextField, { inputTypes } from 'textField'
 import AutoCompleteList from './AutoCompleteListStyle'
+import Divider from 'divider'
+import Loader from 'loader'
 
 type query = string | number | boolean
 export interface AutoCompleteProps {
@@ -50,6 +52,7 @@ export interface AutoCompleteProps {
   readonly required?: boolean
   readonly error?: string | JSX.Element
   readonly selectedItemStatus?: ItemStatus
+  readonly embeddedInSearchForm?: boolean
 }
 
 interface AutoCompleteState {
@@ -103,6 +106,7 @@ export default class AutoComplete extends Component<AutoCompleteProps, AutoCompl
     readOnly: false,
     required: false,
     error: null,
+    embeddedInSearchForm: false,
   }
 
   constructor(props: AutoCompleteProps) {
@@ -266,6 +270,10 @@ export default class AutoComplete extends Component<AutoCompleteProps, AutoCompl
     const listItems = shouldDisplayAutoCompleteList
       ? this.state.items
       : this.props.renderEmptySearch
+    const shouldDisplayDivider =
+      this.props.embeddedInSearchForm && (listItems.length > 0 || shouldDisplayNoResults)
+    const loader =
+      this.props.embeddedInSearchForm && this.state.isSearching ? <Loader size={24} inline /> : null
 
     return (
       <div role="combobox" className={cc([prefix({ autoComplete: true }), this.props.className])}>
@@ -291,7 +299,9 @@ export default class AutoComplete extends Component<AutoCompleteProps, AutoCompl
           required={this.props.required}
           error={this.props.error}
           focusBorder={false}
+          loader={loader}
         />
+        {shouldDisplayDivider && <Divider />}
         {shouldDisplayBusyState && (
           <div className={cc([prefix({ 'autoComplete-body': true }), this.props.bodyClassName])}>
             {this.props.renderBusy({ query: this.state.query })}