+++ src/searchForm/datePicker/section/index.tsx
@@ -1,0 +1,7 @@
+import styled from 'styled-components'
+
+import DatePickerSection from './DatePickerSection'
+
+const StyledDatePickerSection = styled(DatePickerSection)``
+
+export default StyledDatePickerSection